+++ lib/src/dom_builder_base.dart
@@ -55,7 +55,8 @@ abstract class WithValue {
 
 class DOMAttribute implements WithValue {
 
-  static final Set<String> _ATTRIBUTES_VALUE_AS_BOOLEAN = {'checked'};
+  static final Set<String> _ATTRIBUTES_VALUE_AS_BOOLEAN = {'checked','hidden'};
+  static final Set<String> _ATTRIBUTES_VALUE_AS_SET = {'class'};
 
   static final Map<String, Pattern> _ATTRIBUTES_VALUE_AS_LIST_DELIMITERS = {
     'class': ' ',
@@ -76,30 +77,39 @@ class DOMAttribute implements WithValue {
   final String name;
 
   String _value;
-
   List<String> _values;
 
   final String delimiter;
 
-  bool _boolean;
+  bool _valueBoolean;
+  final bool _set;
 
   DOMAttribute(String name,
-      {dynamic value, List values, this.delimiter, dynamic boolean})
+      {dynamic value, List values, this.delimiter, dynamic set, dynamic valueBoolean})
       : name = name.toLowerCase().trim(),
         _value = parseString(value),
         _values = parseListOfStrings(values),
-        _boolean = parseBool(boolean) {
+        _set = parseBool(set,false),
+        _valueBoolean = parseBool(valueBoolean) {
     if (_value != null && _values != null) {
       throw ArgumentError(
           'Attribute $name: Only value or values can be defined, not both.');
     }
-    if (_boolean != null && (_value != null || _values != null)) {
+    if (_valueBoolean != null && (_value != null || _values != null)) {
       throw ArgumentError(
           "Attribute $name: Boolean attribute doesn't have value.");
     }
     if (_values != null && delimiter == null) {
       throw ArgumentError(
-          'Attribute $name: If values is defined a delimiter is required.');
+          'Attribute $name: If values is defined, a delimiter is required.');
+    }
+    if (_value != null && delimiter != null) {
+      throw ArgumentError(
+          'Attribute $name: If value is defined, delimiter should be null.');
+    }
+    if (_set && !isListValue) {
+      throw ArgumentError(
+          'Attribute $name: If is a set, it should be a list value.');
     }
   }
 
@@ -111,15 +121,17 @@ class DOMAttribute implements WithValue {
       _ATTRIBUTES_VALUE_AS_LIST_DELIMITERS_PATTERNS[name];
       assert(delimiterPattern != null);
 
+      var attrSet = _ATTRIBUTES_VALUE_AS_SET.contains(name);
+
       return DOMAttribute(name,
           values: parseListOfStrings(value, delimiterPattern),
-          delimiter: delimiter);
+          delimiter: delimiter, set: attrSet);
     } else {
       var attrBoolean = _ATTRIBUTES_VALUE_AS_BOOLEAN.contains(name);
 
       if (attrBoolean) {
         if (value != null) {
-          return DOMAttribute(name, value: value);
+          return DOMAttribute(name, valueBoolean: value);
         }
       } else {
         return DOMAttribute(name, value: value);
@@ -129,13 +141,14 @@ class DOMAttribute implements WithValue {
     return null ;
   }
 
-  bool get isBoolean => _boolean != null;
+  bool get isBoolean => _valueBoolean != null;
 
   bool get isListValue => delimiter != null;
+  bool get isSet => _set ;
 
   @override
   bool get hasValue {
-    if (isBoolean) return _boolean;
+    if (isBoolean) return _valueBoolean;
 
     if (isListValue) {
       if (isNotEmptyObject(_values)) {
@@ -153,7 +166,7 @@ class DOMAttribute implements WithValue {
 
   @override
   String get value {
-    if (isBoolean) return _boolean.toString();
+    if (isBoolean) return _valueBoolean.toString();
 
     if (isListValue) {
       if (isNotEmptyObject(_values)) {
@@ -171,7 +184,7 @@ class DOMAttribute implements WithValue {
   }
 
   List<String> get values {
-    if (isBoolean) return [_boolean.toString()];
+    if (isBoolean) return [_valueBoolean.toString()];
 
     if (isListValue) {
       if (isNotEmptyObject(_values)) {
@@ -189,7 +202,7 @@ class DOMAttribute implements WithValue {
   bool containsValue(String v) {
     if (isBoolean) {
       v ??= 'false';
-      return _boolean.toString() == v;
+      return _valueBoolean.toString() == v;
     }
 
     if (isListValue) {
@@ -206,7 +219,7 @@ class DOMAttribute implements WithValue {
   }
 
   void setBoolean(dynamic value) {
-    _boolean = parseBool(value, false);
+    _valueBoolean = parseBool(value, false);
   }
 
   void setValue(value) {
@@ -232,17 +245,44 @@ class DOMAttribute implements WithValue {
       return;
     }
 
-    _values ??= [];
-
     var s = parseString(value);
     if (s != null) {
+      _values ??= [];
       _values.add(s);
+
+      if (isSet) {
+        _uniquifyValues();
+      }
+    }
+  }
+
+  void _uniquifyValues() {
+    if (_values == null) return ;
+
+    for (var i = 0 ; i < _values.length;) {
+      var val1 = _values[i] ;
+
+      var duplicated = false ;
+      for (var j = i+1 ; j < _values.length; j++) {
+        var val2 = _values[j] ;
+        if (val1 == val2) {
+          duplicated = true ;
+          break ;
+        }
+      }
+
+      if (duplicated) {
+        _values.removeAt(i) ;
+      }
+      else {
+        i++ ;
+      }
     }
   }
 
   String buildHTML() {
     if (isBoolean) {
-      return _boolean ? name : '';
+      return _valueBoolean ? name : '';
     }
 
     var htmlValue = value;
@@ -258,7 +298,7 @@ class DOMAttribute implements WithValue {
 
   @override
   String toString() {
-    return 'DOMAttribute{name: $name, _value: $_value, _values: $_values, delimiter: $delimiter, _boolean: $_boolean}';
+    return 'DOMAttribute{name: $name, _value: $_value, _values: $_values, delimiter: $delimiter, _valueBoolean: $_valueBoolean}';
   }
 
 }
@@ -1479,6 +1519,18 @@ class DOMElement extends DOMNode {
 
   /// Returns the attribute `class`.
   String get classes => getAttributeValue('class');
+
+  /// Returns the list of class names of the attribute `class`.
+  List<String> get classesList {
+    var attribute = getAttribute('class');
+    if ( attribute == null ) return [] ;
+    return attribute.values ?? [] ;
+  }
+
+  void addClass(String className) {
+    appendToAttribute('class', className) ;
+  }
+
 
   /// Returns the attribute `style`.
   String get style => getAttributeValue('style');